Finish migrating tests.
